Refactor GenericVectorBaseMap into separate modules
